Move NativeAllocatorPoolStats and SPI from arrow-spi to server
Drops the api project(":libs:opensearch-arrow-spi") dependency from
server/build.gradle by relocating NativeAllocatorPoolStats and the
ArrowAllocatorPlugin SPI marker into server. Mirrors the placement
of AnalyticsBackendNativeMemoryStats / SearchBackEndPlugin (the
NativeMemory PR pattern).

Server is now arrow-spi-free.

Migration:
- NativeAllocatorPoolStats:
    libs/arrow-spi → server/src/main/java/org/opensearch/plugin/stats/
- ArrowAllocatorPlugin SPI:
    libs/arrow-spi → server/src/main/java/org/opensearch/plugins/
- SPI shape changes from `getNativeAllocator() -> NativeAllocator` to
  `getNativeAllocatorStatsSupplier() -> Supplier<NativeAllocatorPoolStats>`,
  matching SearchBackEndPlugin.getAnalyticsBackendNativeMemoryStats.
- NativeAllocator.stats() removed from arrow-spi; ArrowNativeAllocator
  retains the method but now returns the server-resident type.
- ArrowBasePlugin implements the new server-side SPI; the supplier
  closes over the live allocator and produces a fresh snapshot per
  invocation.
- NodeService swaps `setNativeAllocator(NativeAllocator)` for
  `setNativeAllocatorStatsSupplier(Supplier<...>)`. End-to-end response
  shape is unchanged.
- Drops the libs/arrow-spi/build.gradle jar-hell exclude (no longer
  needed since server doesn't pull arrow-spi transitively).
