+++ server/build.gradle
@@ -73,7 +73,6 @@ dependencies {
   api project(":libs:opensearch-geo")
   api project(":libs:opensearch-telemetry")
   api project(":libs:opensearch-task-commons")
-  api project(":libs:opensearch-arrow-spi")
 
 
   compileOnly project(":libs:agent-sm:bootstrap")
