+++ server/src/test/java/org/opensearch/plugin/stats/NativeAllocatorPoolStatsTests.java
@@ -6,7 +6,7 @@
  * compatible open source license.
  */
 
-package org.opensearch.arrow.spi;
+package org.opensearch.plugin.stats;
 
 import org.opensearch.common.io.stream.BytesStreamOutput;
 import org.opensearch.common.xcontent.json.JsonXContent;
